CPT, taxonomies, sidebars are separate modules now
